pyproject.toml
[project]
name = "ci-failure-classifier"
version = "0.1.0"
description = "Classify CI failures as Infrastructure vs Playwright test failures"
readme = "how-to-infra-vs-playwright.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.0.0",
]

[dependency-groups]
dev = []

